requirements.txt
pytesseract
opencv-python
Pillow
pandas
openpyxl
python-dotenv
dotenv
google-cloud-vision
openpyxl 
pandas
easyocr
